Add a lightweight Ethereum wallet for testing reverse direction (#43)
Add a small example of an Ethereum wallet that subscribes to the starbridge network.

As we start implementing transfers going the other way we need to have something simple for testing the transfer going the other way, with a wallet picking up the transaction and submitting it to Ethereum.

This tool is really temporary and throw-away, because ultimately we'll have a dapp that a wallet will interact with and that will surface transfers that the wallet can submit to the contract. We should expect this code to be deleted, but it is useful in testing locally until we figure the dapp part out.

The example doesn't yet post the seen message to the contract and that will be completed in another change. There are TODOs introduced at some key locations where that will take place.