Prototype of Ethereum - Stellar inverse transaction flow (#50)
This commit adds initial versions of different services of Starbridge server. In
general the server is currenly divided into two parts: frontend and backend.
Frontend handles HTTP requests and allows users to create `SignatureRequest`'s
(requests to sign an inverse transaction) and send it to a backend server via a
queue (or a DB). Backend server is responsible for streaming blockchain data,
handling `SignatureRequest`'s and finally creating and signing transactions.
This separation was made to allow more secure architectures in which frontend
server does not have access to signing part of the app.

In the backend server, transaction builder and transaction signer are separate
entities. This is done to (maybe) support using secure enclaves (like AWS Nitro
Enclaves) that will store the secret key and sign transactions.

List of services:
* `backend/Worker` - Main backend worker handling `SignatureRequest`
* `stellar/controllers/StellarGetInverseTransactionForEthereum` - Controller
  responsible for building inverse transactions, contains simple logic for
  checking if a new `SignatureRequest` should be created or if the inverse
  transaction should be returned to the user
* `stellar/signer/Signer` - Stellar transaction signer
* `stellar/txbuilder/Builder` - Stellar transaction builder, currently ETH tx
  hardcoded
* `stellar/txobserver/Observer` - Observer of Stellar transactions, currently
  using testnet. 
* `store/Memory` - Storage system implemented in memory, in the future: Postgres
  DB.

The current version contains at least one race condition because
`UpsertOutgoingStellarTransaction` is called concurrently from two different
services. This should be solved by `SELECT ... FOR UPDATE` in Postgres but I
wonder if we should allow just one DB transaction at a time for security reasons
(to decrease number of synchronization issues).