Move Stellar tx observer to main backend worker (#52)
This commit moves Stellar tx observer to be run in the same go routine as
backend worker. Both the worker and observer update the outgoing Stellar
transactions state in a DB. Doing this in two go routines adds complexity to
prevent race conditions in the code. In the new code the worker will always call
observer to process all the latest ledgers before processing signing requests.