Integration test framework and a simple test (#53)
This commit adds an integration test framework for testing Starbridge in a
standalone network and a simple integration test. The integration test is a
happy path test: it starts three Starbridge validators, configure the Stellar
account and ask each of them for signatures for the inverse transaction. When
signatures are received they are concatenated and submitted to the network.