Change store to use Postgres DB (#51)
In 8533df3 the initial implementation of data store was using in memory
structures because I wasn't sure if we will use Postgres or a DB engine that
requires less resources to run. After consideration I think we should use
Postgres for multiple reasons. It's the main relational DB used in SDF, we have
expertise in running it. It contains the ACL features that can be useful when
Starbridge is deployed as frontend and backend servers. Finally, it's easy to
recover the DB in case of failures.